layer-tools/websearch-orch/requirements-dspy-pilot.txt
dspy-ai>=2.6.0
jsonschema>=4.23.0
